requirements.txt
amqp==2.2.2
argh==0.26.2
astroid==1.6.4
autopep8==1.3.5
billiard==3.5.0.3
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-bootstrap3==10.0.1
django-filter==1.1.0
django-suit==0.2.26
django-widget-tweaks==1.4.2
idna==2.6
isort==4.3.4
kombu==4.1.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
pathtools==0.1.2
pkg-resources==0.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycodestyle==2.4.0
pylint==1.9.1
pytz==2018.4
PyYAML==3.12
redis==2.10.6
requests==2.18.4
six==1.11.0
urllib3==1.22
uWSGI==2.0.17
vine==1.1.4
watchdog==0.8.3
wrapt==1.10.11
